chore: ignore VS cache and untrack .vs/TestResults
