+++ PasswordStrengthDemo/Program.cs
@@ -1,3 +1,3 @@
 ﻿using PasswordStrengthLib;
 
-Console.WriteLine(PasswordEvaluator.Evaluate("Ab1!")); // should print STRONG
+Console.WriteLine(PasswordEvaluator.Evaluate("Ab1!xxxx"));
