+++ PasswordStrengthSolution.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.14.36414.22
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PasswordStrengthLib", "PasswordStrengthLib\PasswordStrengthLib.csproj", "{ECEF1D0A-3ABB-40D9-9BB6-305E176210FB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PasswordStrengthDemo", "PasswordStrengthDemo\PasswordStrengthDemo.csproj", "{6FBDB8F2-216D-4AA9-B09D-77BD9D9B3997}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PasswordStrengthLib.Tests", "PasswordStrengthLib.Tests\PasswordStrengthLib.Tests.csproj", "{E7C7FF87-9CEE-4EE0-B1A9-6BD8C1FF59AC}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{8EC462FD-D22E-90A8-E5CE-7E832BA40C5D}"
@@ -24,10 +22,6 @@ Global
 		{ECEF1D0A-3ABB-40D9-9BB6-305E176210FB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ECEF1D0A-3ABB-40D9-9BB6-305E176210FB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{ECEF1D0A-3ABB-40D9-9BB6-305E176210FB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6FBDB8F2-216D-4AA9-B09D-77BD9D9B3997}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6FBDB8F2-216D-4AA9-B09D-77BD9D9B3997}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6FBDB8F2-216D-4AA9-B09D-77BD9D9B3997}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6FBDB8F2-216D-4AA9-B09D-77BD9D9B3997}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E7C7FF87-9CEE-4EE0-B1A9-6BD8C1FF59AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E7C7FF87-9CEE-4EE0-B1A9-6BD8C1FF59AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E7C7FF87-9CEE-4EE0-B1A9-6BD8C1FF59AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
